# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export type RequestUploadUrlResponseMetadata = {
  name: string;
  size: number;
  contentType: string;
};

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata: RequestUploadUrlResponseMetadata;
}

export type MediaItemType = typeof MediaItemType[keyof typeof MediaItemType];


export const MediaItemType = {
  image: 'image',
  video: 'video',
} as const;

export type MediaItemVisibility = typeof MediaItemVisibility[keyof typeof MediaItemVisibility];


export const MediaItemVisibility = {
  public: 'public',
  private: 'private',
} as const;

export interface MediaItem {
  id: string;
  objectPath: string;
  type: MediaItemType;
  caption?: string | null;
  uploader?: string | null;
  visibility: MediaItemVisibility;
  createdAt: string;
}

export type CreateMediaBodyType = typeof CreateMediaBodyType[keyof typeof CreateMediaBodyType];


export const CreateMediaBodyType = {
  image: 'image',
  video: 'video',
} as const;

export type CreateMediaBodyVisibility = typeof CreateMediaBodyVisibility[keyof typeof CreateMediaBodyVisibility];


export const CreateMediaBodyVisibility = {
  public: 'public',
  private: 'private',
} as const;

export interface CreateMediaBody {
  objectPath: string;
  type: CreateMediaBodyType;
  caption?: string | null;
  uploader?: string | null;
  visibility: CreateMediaBodyVisibility;
}

export type UpdateMediaBodyVisibility = typeof UpdateMediaBodyVisibility[keyof typeof UpdateMediaBodyVisibility];


export const UpdateMediaBodyVisibility = {
  public: 'public',
  private: 'private',
} as const;

export interface UpdateMediaBody {
  visibility?: UpdateMediaBodyVisibility;
  caption?: string | null;
}