# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Request a presigned upload URL
 */
export const RequestUploadUrlBody = zod.object({
  "name": zod.string(),
  "size": zod.number(),
  "contentType": zod.string()
})

export const RequestUploadUrlResponse = zod.object({
  "uploadURL": zod.string(),
  "objectPath": zod.string(),
  "metadata": zod.object({
  "name": zod.string(),
  "size": zod.number(),
  "contentType": zod.string()
})
})


export const GetObjectParams = zod.object({
  "objectPath": zod.coerce.string()
})


/**
 * @summary List all media items
 */
export const ListMediaResponseItem = zod.object({
  "id": zod.string(),
  "objectPath": zod.string(),
  "type": zod.enum(['image', 'video']),
  "caption": zod.string().nullish(),
  "uploader": zod.string().nullish(),
  "visibility": zod.enum(['public', 'private']),
  "createdAt": zod.string()
})
export const ListMediaResponse = zod.array(ListMediaResponseItem)


/**
 * @summary Create a media record
 */
export const CreateMediaBody = zod.object({
  "objectPath": zod.string(),
  "type": zod.enum(['image', 'video']),
  "caption": zod.string().nullish(),
  "uploader": zod.string().nullish(),
  "visibility": zod.enum(['public', 'private'])
})


/**
 * @summary Update a media item
 */
export const UpdateMediaParams = zod.object({
  "id": zod.coerce.string()
})

export const UpdateMediaBody = zod.object({
  "visibility": zod.enum(['public', 'private']).optional(),
  "caption": zod.string().nullish()
})

export const UpdateMediaResponse = zod.object({
  "id": zod.string(),
  "objectPath": zod.string(),
  "type": zod.enum(['image', 'video']),
  "caption": zod.string().nullish(),
  "uploader": zod.string().nullish(),
  "visibility": zod.enum(['public', 'private']),
  "createdAt": zod.string()
})


/**
 * @summary Delete a media item
 */
export const DeleteMediaParams = zod.object({
  "id": zod.coerce.string()
})